fixes layers
