+++ backend/src/domain/value_objects.py
@@ -1,0 +1,6 @@
+from cryptography.utils import Enum
+
+
+class RecordStatus(Enum):
+    OPEN = "OPEN"
+    CLOSED = "CLOSED"
